annotate Morlet1DSpec inner constructor

+++ src/morlet1d.jl
@@ -6,9 +6,9 @@ immutable Morlet1DSpec{T<:Number} <: Abstract1DSpec{T}
     max_scale # ::realtype(T) (imposed by inner constructor)
     nFilters_per_octave::Int
     nOctaves::Int
-    function Morlet1DSpec(ɛ, signaltype::Type{T}, log2_length,
+    function Morlet1DSpec(ɛ, signaltype::Type{T}, log2_length::Int,
                                      max_qualityfactor, max_scale,
-                                     nFilters_per_octave, nOctaves)
+                                     nFilters_per_octave::Int, nOctaves::Int)
         RealT = realtype(T)
         new(RealT(ɛ), T, log2_length, RealT(max_qualityfactor),
             RealT(max_scale), nFilters_per_octave, nOctaves)
